Ensuring all edits are flake8 compliant
