+++ PhageScanner/main/orffinder_wrappers.py
@@ -6,12 +6,13 @@ Description:
     Phanotate, since it is focued on bacteriophages, but other ORF
     finding tools can be added for other species.
 """
+
 import logging
 import os
+import re
 from abc import ABC, abstractmethod
 from enum import Enum
 from pathlib import Path
-import re
 
 from PhageScanner.main.exceptions import IncorrectYamlError, MissingFileError
 from PhageScanner.main.utils import CommandLineUtils
@@ -94,9 +95,10 @@ class PhanotateWrapper(OrfFinderWrapper):
     @staticmethod
     def get_info_from_name(fasta_entry_name: str):
         """Get information from the fasta entry name."""
-
-        # Pattern expects Phanotate output of "NC_022762.1_CDS_[4760..5803] [note=score:-1.440691E+08]"
-        pattern = r"(\w+\.\d+)_CDS_\[(\d+)\.\.(\d+)\] \[note=score:(-?\d+\.\d+E[+-]\d+)\]"
+        # Pattern expects "NC_022762.1_CDS_[4760..5803] [note=score:-1.440691E+08]"
+        pattern = (
+            r"(\w+\.\d+)_CDS_\[(\d+)\.\.(\d+)\] \[note=score:(-?\d+\.\d+E[+-]\d+)\]"
+        )
         logging.info(fasta_entry_name)
         match = re.search(pattern, fasta_entry_name)
