+++ PhageScanner/tests/test_DNA.py
@@ -4,6 +4,7 @@ Testng DNA class.
 This set of unit tests tests the DNA
 class for expected functionality.
 """
+
 # Testing
 import unittest
